mqtt: Move handle mappings out of RobotMqttHandle.js
