+++ lib/mqtt/handles/RobotMqttHandle.js
@@ -1,12 +1,10 @@
 const CallbackAttributeSubscriber = require("../../entities/CallbackAttributeSubscriber");
 const MqttHandle = require("./MqttHandle");
 
-const capabilities = require("../../core/capabilities");
-const capabilityHandles = require("../capabilities");
 const MapNodeMqttHandle = require("./MapNodeMqttHandle");
-const stateAttrs = require("../../entities/state/attributes");
-const stateHandles = require("../status");
 const VacuumHassComponent = require("../homeassistant/components/VacuumHassComponent");
+const {CAPABILITY_TYPE_TO_HANDLE_MAPPING} = require("./HandleMappings");
+const {STATUS_ATTR_TO_HANDLE_MAPPING} = require("./HandleMappings");
 
 /**
  * This class represents the robot as a Homie device
@@ -139,27 +137,5 @@ class RobotMqttHandle extends MqttHandle {
         return this.mapHandle;
     }
 }
-
-const CAPABILITY_TYPE_TO_HANDLE_MAPPING = {
-    [capabilities.BasicControlCapability.TYPE]: capabilityHandles.BasicControlCapabilityMqttHandle,
-    [capabilities.FanSpeedControlCapability.TYPE]: capabilityHandles.IntensityPresetCapabilityMqttHandle,
-    [capabilities.LocateCapability.TYPE]: capabilityHandles.LocateCapabilityMqttHandle,
-    [capabilities.WaterUsageControlCapability.TYPE]: capabilityHandles.IntensityPresetCapabilityMqttHandle,
-};
-
-const STATUS_ATTR_TO_HANDLE_MAPPING = [
-    {
-        matcher: {attributeClass: stateAttrs.AttachmentStateAttribute.name},
-        handle: stateHandles.AttachmentStateMqttHandle
-    },
-    {
-        matcher: {attributeClass: stateAttrs.BatteryStateAttribute.name},
-        handle: stateHandles.BatteryStateMqttHandle
-    },
-    {
-        matcher: {attributeClass: stateAttrs.StatusStateAttribute.name},
-        handle: stateHandles.StatusStateMqttHandle
-    }
-];
 
 module.exports = RobotMqttHandle;
